+++ modules/app/index.php
@@ -30,7 +30,20 @@
 				$decode = json_decode($read,true);				
 				?>
 
-				<?php foreach ($decode as $comments): ?>
+				<div class="alert alert-primary" role="alert">
+					<div class="bg-light p-2 border-radius mt-2">
+						<h5 class="card-title">
+							<?php echo $decode['post']['header']; ?>
+						</h5>
+						<p class="card-text">
+							<?php echo $decode['post']['message']; ?>
+							<br/>
+							<?php echo str_replace('href="/', 'href="https://mbasic.facebook.com/', $decode['post']['media']); ?>							
+						</p>
+					</div>
+				</div>
+
+				<?php foreach ($decode['comments'] as $comments): ?>
 					<div class="card text-dark mb-2">
 						<div class="card-body">
 							<h5 class="card-title">
@@ -39,7 +52,7 @@
 							<p class="card-text">
 								<?php echo $comments['message']; ?>
 								<br/>
-								<?php echo $comments['media']; ?>
+								<?php echo str_replace('href="/', 'href="https://mbasic.facebook.com/', $comments['media']); ?>								
 							</p>
 
 							<?php if ($comments['reply']): ?>
@@ -52,7 +65,7 @@
 											<p class="card-text">
 												<?php echo $commentsR['message']; ?>
 												<br/>
-												<?php echo $commentsR['media']; ?>
+												<?php echo str_replace('href="/', 'href="https://mbasic.facebook.com/', $commentsR['media']);?>												
 											</p>
 										</div>
 									</div>