+++ modules/history/read.php
@@ -10,9 +10,20 @@ $decode = json_decode($read,true);
 
 			<div class="alert alert-primary" role="alert">
 				Read history <a target="_blank" href="https://facebook.com/<?php echo explode('.', $file)[0] ?>"><?php echo $file ?></a>
+				<div class="bg-light p-2 border-radius mt-2">
+					<h5 class="card-title">
+						<?php echo $decode['post']['header']; ?>
+					</h5>
+					<p class="card-text">
+						<?php echo $decode['post']['message']; ?>
+						<br/>
+						<?php echo str_replace('href="/', 'href="https://mbasic.facebook.com/', $decode['post']['media']); 
+						?>
+					</p>
+				</div>
 			</div>
 
-			<?php foreach ($decode as $comments): ?>
+			<?php foreach ($decode['comments'] as $comments): ?>
 				<div class="card text-dark mb-2">
 					<div class="card-body">
 						<h5 class="card-title">
@@ -21,7 +32,8 @@ $decode = json_decode($read,true);
 						<p class="card-text">
 							<?php echo $comments['message']; ?>
 							<br/>
-							<?php echo $comments['media']; ?>
+							<?php echo str_replace('href="/', 'href="https://mbasic.facebook.com/', $comments['media']);
+							?> 
 						</p>
 
 						<?php if ($comments['reply']): ?>
@@ -34,7 +46,8 @@ $decode = json_decode($read,true);
 										<p class="card-text">
 											<?php echo $commentsR['message']; ?>
 											<br/>
-											<?php echo $commentsR['media']; ?>
+											<?php echo str_replace('href="/', 'href="https://mbasic.facebook.com/', $commentsR['media']);
+											?> 
 										</p>
 									</div>
 								</div>
@@ -46,7 +59,7 @@ $decode = json_decode($read,true);
 
 			<div class="alert alert-primary" role="alert">
 				<a href="./?module=history" class="alert-link">
-					 &lt; Back
+					&lt; Back
 				</a>
 			</div>