+++ modules/logout.php
@@ -1,4 +1,10 @@
 <?php defined('BASEPATH') OR exit('no direct script access allowed');
 
-session_destroy();
+
+// remove session
+unset($_SESSION['login']);
+
+// remove cookie
+setcookie('login', false, time()-1000, '/');
+
 header("location: ./");