+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ## Change Log
+**18 Februari 2022**
+- Update library
+- Comment-reply fix + add delay
+- Cookie login
+- Add previous login
+- Add post info
+
 **17 Februari 2022**
 - Release